+++ viadot/flows/adls_to_azure_sql.py
@@ -3,7 +3,6 @@ import os
 from typing import Any, Dict, List, Literal
 
 import pandas as pd
-import numpy as np
 from prefect import Flow, task
 from prefect.backend import get_key_value
 from prefect.utilities import logging