+++ viadot/tasks/__init__.py
@@ -44,3 +44,7 @@ from .duckdb import DuckDBCreateTableFromParquet, DuckDBQuery, DuckDBToDF
 from .sql_server import SQLServerCreateTable, SQLServerToDF, SQLServerQuery
 
 from .epicor import EpicorOrdersToDF
+
+from .sftp import SftpToDF, SftpList
+from .mindful import MindfulToCSV
+